Initial commit for theme update manager logic
* ENHANCEMENT: Adds support to let themes update within the Update Manager Add On.

This includes the filter `pmproum_owned_themes` that we can hook into from a theme's side to automatically add the ability to serve update via the Update Manager plugin. Themes should still bundle this code natively or require the Update Manager Plugin.

If a theme bundles update logic code inside their theme, it would be best to check if the Update Manager is installed or not and let the Update Manager handle the updates (to avoid duplicate calls to the JSON license server).